Simplify JSArrayUtils
