+++ jso/apis/src/main/java/org/teavm/jso/core/JSString.java
@@ -35,8 +35,17 @@ public abstract class JSString implements JSObject, BlobConvertible {
     @JSBody(params = "str", script = "return str;")
     private static native String stringValue(JSString str);
 
-    @JSBody(params = "str", script = "return str;")
+    @JSBody(params = "val", script = "return str;")
     public static native JSString valueOf(String str);
+
+    @JSBody(params = "val", script = "return val.toString()")
+    public static native JSString valueOf(int val);
+
+    @JSBody(params = "val", script = "return val.toString()")
+    public static native JSString valueOf(double val);
+
+    @JSBody(params = "val", script = "return val.toString()")
+    public static native JSString valueOf(float val);
 
     @JSBody(params = "code", script = "return String.fromCharCode(code)")
     public static native JSString fromCharCode(int code);
