+++ jso/apis/src/main/java/org/teavm/jso/core/utils/JSMapFunction.java
@@ -1,0 +1,32 @@
+/*
+ *  Copyright 2015 Jan-Felix Wittmann.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.teavm.jso.core.utils;
+
+import org.teavm.jso.JSFunctor;
+import org.teavm.jso.JSObject;
+import org.teavm.jso.core.JSArray;
+
+/**
+*
+* @author Jan-Felix Wittmann
+*/
+@JSFunctor
+@FunctionalInterface
+public interface JSMapFunction<T extends JSObject, R extends JSObject> extends JSObject {
+
+    R apply(T element, int index, JSArray<T> arr);
+
+}
